# Benchmark for DiscreteTrajectory