--- a/benchmarks/discrete_trajectory.cpp
+++ b/benchmarks/discrete_trajectory.cpp
@@ -1,21 +1,27 @@
-
+﻿
 // .\Release\x64\benchmarks.exe --benchmark_filter=DiscreteTrajectory
 
 #include "physics/discrete_trajectory.hpp"
 
+#include "astronomy/epoch.hpp"
 #include "base/not_null.hpp"
 #include "benchmark/benchmark.h"
+#include "geometry/barycentre_calculator.hpp"
 #include "geometry/frame.hpp"
 #include "ksp_plugin/frames.hpp"
+#include "physics/discrete_trajectory_types.hpp"
 #include "quantities/elementary_functions.hpp"
 #include "quantities/named_quantities.hpp"
 #include "quantities/quantities.hpp"
+#include "testing_utilities/discrete_trajectory_factories.hpp"
 
 namespace principia {
 namespace physics {
 
+using astronomy::InfiniteFuture;
 using base::make_not_null_unique;
 using base::not_null;
+using geometry::Barycentre;
 using geometry::Displacement;
 using geometry::Frame;
 using geometry::Handedness;
@@ -23,6 +29,7 @@ using geometry::Inertial;
 using geometry::Instant;
 using geometry::Velocity;
 using ksp_plugin::World;
+using physics::internal_discrete_trajectory_types::Timeline;
 using quantities::AngularFrequency;
 using quantities::Cos;
 using quantities::Sin;
@@ -30,136 +37,154 @@ using quantities::Time;
 using quantities::si::Metre;
 using quantities::si::Radian;
 using quantities::si::Second;
+using testing_utilities::NewCircularTrajectoryTimeline;
+using testing_utilities::NewMotionlessTrajectoryTimeline;
 
 namespace {
 
-// Creates a motionless trajectory with the given number of steps.
-not_null<std::unique_ptr<DiscreteTrajectory<World>>> CreateMotionlessTrajectory(
-    int const steps) {
-  auto trajectory = make_not_null_unique<DiscreteTrajectory<World>>();
-  for (Time t = 0 * Second; t < steps * Second; t += 1 * Second) {
-    trajectory->Append(Instant() + t, {World::origin, World::unmoving});
+// Constructs a trajectory by assigning the points in |timeline| to segments
+// defined by |splits|, which must be doubles in [0, 1].
+DiscreteTrajectory<World> MakeTrajectory(Timeline<World> const& timeline,
+                                         std::vector<double> const& splits) {
+  DiscreteTrajectory<World> trajectory;
+  Instant const t_min = timeline.begin()->time;
+  Instant const t_max = timeline.rbegin()->time;
+  auto it_split = splits.begin();
+  std::optional<Instant> t_split;
+  for (auto const& [t, degrees_of_freedom] : timeline) {
+    // The computation of |t_split| is complicated enough that we want to do it
+    // at a single place.
+    if (!t_split.has_value()) {
+      if (it_split == splits.end()) {
+        t_split = InfiniteFuture;
+      } else {
+        double const split = *it_split;
+        CHECK_LE(0, split);
+        CHECK_LE(split, 1);
+        t_split = Barycentre<Instant, double>({t_min, t_max},
+                                              {1 - split, split});
+      }
+    }
+    if (t >= t_split) {
+      ++it_split;
+      t_split = std::nullopt;
+      trajectory.NewSegment();
+    }
+    trajectory.Append(t, degrees_of_freedom);
   }
   return trajectory;
-}
-
-// Creates a circular trajectory with the given number of steps in the XZ plane.
-not_null<std::unique_ptr<DiscreteTrajectory<World>>> CreateCircularTrajectory(
-    int const steps) {
-  auto trajectory = make_not_null_unique<DiscreteTrajectory<World>>();
-  constexpr AngularFrequency ω = 1 * Radian / Second;
-  for (Time t = 0 * Second; t < steps * Second; t += 1 * Second) {
-    double const sin_ωt = Sin(ω * t);
-    double const cos_ωt = Cos(ω * t);
-    trajectory->Append(
-        Instant() + t,
-        {World::origin +
-             Displacement<World>({cos_ωt * Metre, 0 * Metre, sin_ωt * Metre}),
-         Velocity<World>({-sin_ωt * Metre / Second,
-                          0 * Metre / Second,
-                          cos_ωt * Metre / Second})});
-  }
-  return trajectory;
-}
-
-// Forks |parent| at a position |pos| of the way through.
-// |parent| should be nonempty.
-// |pos| should be in [0, 1].
-not_null<DiscreteTrajectory<World>*> ForkAt(DiscreteTrajectory<World>& parent,
-                                            double const pos) {
-  CHECK(!parent.Empty());
-  CHECK_GE(pos, 0);
-  CHECK_LE(pos, 1);
-  Instant const desired_fork_time =
-      parent.t_min() + (parent.t_max() - parent.t_min()) * pos;
-  auto const fork_it = parent.LowerBound(desired_fork_time);
-  return parent.NewForkWithCopy(fork_it->time);
 }
 
 }  // namespace
 
 void BM_DiscreteTrajectoryFront(benchmark::State& state) {
-  not_null<std::unique_ptr<DiscreteTrajectory<World>>> const trajectory =
-      CreateMotionlessTrajectory(4);
-  not_null<DiscreteTrajectory<World>*> const fork =
-      ForkAt(*ForkAt(*trajectory, 0.5), 0.75);
+  Instant const t0;
+  auto const timeline = NewMotionlessTrajectoryTimeline(World::origin,
+                                                        /*Δt=*/1 * Second,
+                                                        /*t1=*/t0,
+                                                        /*t2=*/t0 + 4 * Second);
+  auto const trajectory = MakeTrajectory(timeline, {0.5, 0.75});
+  auto const& segment = trajectory.segments().back();
 
   for (auto _ : state) {
-    fork->front();
+    segment.front();
   }
 }
-
 void BM_DiscreteTrajectoryBack(benchmark::State& state) {
-  not_null<std::unique_ptr<DiscreteTrajectory<World>>> const trajectory =
-      CreateMotionlessTrajectory(4);
-  not_null<DiscreteTrajectory<World>*> const fork =
-      ForkAt(*ForkAt(*trajectory, 0.5), 0.75);
+  Instant const t0;
+  auto const timeline = NewMotionlessTrajectoryTimeline(World::origin,
+                                                        /*Δt=*/1 * Second,
+                                                        /*t1=*/t0,
+                                                        /*t2=*/t0 + 4 * Second);
+  auto const trajectory = MakeTrajectory(timeline, {0.5, 0.75});
+  auto const& segment = trajectory.segments().back();
 
   for (auto _ : state) {
-    fork->back();
+    segment.back();
   }
 }
 
 void BM_DiscreteTrajectoryBegin(benchmark::State& state) {
-  not_null<std::unique_ptr<DiscreteTrajectory<World>>> const trajectory =
-      CreateMotionlessTrajectory(4);
-  not_null<DiscreteTrajectory<World>*> const fork =
-      ForkAt(*ForkAt(*trajectory, 0.5), 0.75);
+  Instant const t0;
+  auto const timeline = NewMotionlessTrajectoryTimeline(World::origin,
+                                                        /*Δt=*/1 * Second,
+                                                        /*t1=*/t0,
+                                                        /*t2=*/t0 + 4 * Second);
+  auto const trajectory = MakeTrajectory(timeline, {0.5, 0.75});
+  auto const& segment = trajectory.segments().back();
 
   for (auto _ : state) {
-    fork->begin();
+    segment.begin();
   }
 }
 
 void BM_DiscreteTrajectoryEnd(benchmark::State& state) {
-  not_null<std::unique_ptr<DiscreteTrajectory<World>>> const trajectory =
-      CreateMotionlessTrajectory(4);
-  not_null<DiscreteTrajectory<World>*> const fork =
-      ForkAt(*ForkAt(*trajectory, 0.5), 0.75);
+  Instant const t0;
+  auto const timeline = NewMotionlessTrajectoryTimeline(World::origin,
+                                                        /*Δt=*/1 * Second,
+                                                        /*t1=*/t0,
+                                                        /*t2=*/t0 + 4 * Second);
+  auto const trajectory = MakeTrajectory(timeline, {0.5, 0.75});
+  auto const& segment = trajectory.segments().back();
 
   for (auto _ : state) {
-    fork->end();
+    segment.end();
   }
 }
 
 void BM_DiscreteTrajectoryTMin(benchmark::State& state) {
-  not_null<std::unique_ptr<DiscreteTrajectory<World>>> const trajectory =
-      CreateMotionlessTrajectory(4);
-  not_null<DiscreteTrajectory<World>*> const fork =
-      ForkAt(*ForkAt(*trajectory, 0.5), 0.75);
+  Instant const t0;
+  auto const timeline = NewMotionlessTrajectoryTimeline(World::origin,
+                                                        /*Δt=*/1 * Second,
+                                                        /*t1=*/t0,
+                                                        /*t2=*/t0 + 4 * Second);
+  auto const trajectory = MakeTrajectory(timeline, {0.5, 0.75});
+  auto const& segment = trajectory.segments().back();
 
   for (auto _ : state) {
-    fork->t_min();
+    segment.t_min();
   }
 }
 
 void BM_DiscreteTrajectoryTMax(benchmark::State& state) {
-  not_null<std::unique_ptr<DiscreteTrajectory<World>>> const trajectory =
-      CreateMotionlessTrajectory(4);
-  not_null<DiscreteTrajectory<World>*> const fork =
-      ForkAt(*ForkAt(*trajectory, 0.5), 0.75);
+  Instant const t0;
+  auto const timeline = NewMotionlessTrajectoryTimeline(World::origin,
+                                                        /*Δt=*/1 * Second,
+                                                        /*t1=*/t0,
+                                                        /*t2=*/t0 + 4 * Second);
+  auto const trajectory = MakeTrajectory(timeline, {0.5, 0.75});
+  auto const& segment = trajectory.segments().back();
 
   for (auto _ : state) {
-    fork->t_max();
+    segment.t_max();
   }
 }
 
 void BM_DiscreteTrajectoryCreateDestroy(benchmark::State& state) {
+  Instant const t0;
   int const steps = state.range(0);
+  auto const timeline =
+      NewMotionlessTrajectoryTimeline(World::origin,
+                                      /*Δt=*/1 * Second,
+                                      /*t1=*/t0,
+                                      /*t2=*/t0 + steps * Second);
   for (auto _ : state) {
-    CreateMotionlessTrajectory(steps);
+    MakeTrajectory(timeline, {0.5, 0.75});
   }
 }
 
 void BM_DiscreteTrajectoryIterate(benchmark::State& state) {
+  Instant const t0;
   int const steps = state.range(0);
-  not_null<std::unique_ptr<DiscreteTrajectory<World>>> const trajectory =
-      CreateMotionlessTrajectory(steps);
-  not_null<DiscreteTrajectory<World>*> const fork =
-      ForkAt(*ForkAt(*trajectory, 0.5), 0.75);
+  auto const timeline =
+      NewMotionlessTrajectoryTimeline(World::origin,
+                                      /*Δt=*/1 * Second,
+                                      /*t1=*/t0,
+                                      /*t2=*/t0 + steps * Second);
+  auto const trajectory = MakeTrajectory(timeline, {0.5, 0.75});
 
-  auto const begin = fork->begin();
-  auto const end = fork->end();
+  auto const begin = trajectory.begin();
+  auto const end = trajectory.end();
   for (auto _ : state) {
     for (auto it = begin; it != end; ++it) {
     }
@@ -167,14 +192,17 @@ void BM_DiscreteTrajectoryIterate(benchmark::State& state) {
 }
 
 void BM_DiscreteTrajectoryReverseIterate(benchmark::State& state) {
+  Instant const t0;
   int const steps = state.range(0);
-  not_null<std::unique_ptr<DiscreteTrajectory<World>>> const trajectory =
-      CreateMotionlessTrajectory(steps);
-  not_null<DiscreteTrajectory<World>*> const fork =
-      ForkAt(*ForkAt(*trajectory, 0.5), 0.75);
+  auto const timeline =
+      NewMotionlessTrajectoryTimeline(World::origin,
+                                      /*Δt=*/1 * Second,
+                                      /*t1=*/t0,
+                                      /*t2=*/t0 + steps * Second);
+  auto const trajectory = MakeTrajectory(timeline, {0.5, 0.75});
 
-  auto const begin = fork->begin();
-  auto const end = fork->end();
+  auto const begin = trajectory.begin();
+  auto const end = trajectory.end();
   for (auto _ : state) {
     for (auto it = end; it != begin; --it) {
     }
@@ -182,55 +210,72 @@ void BM_DiscreteTrajectoryReverseIterate(benchmark::State& state) {
 }
 
 void BM_DiscreteTrajectoryFind(benchmark::State& state) {
+  Instant const t0;
   int const steps = state.range(0);
-  not_null<std::unique_ptr<DiscreteTrajectory<World>>> const trajectory =
-      CreateMotionlessTrajectory(steps);
-  not_null<DiscreteTrajectory<World>*> const fork =
-      ForkAt(*ForkAt(*trajectory, 0.5), 0.75);
+  auto const timeline =
+      NewMotionlessTrajectoryTimeline(World::origin,
+                                      /*Δt=*/1 * Second,
+                                      /*t1=*/t0,
+                                      /*t2=*/t0 + steps * Second);
+  auto const trajectory = MakeTrajectory(timeline, {0.5, 0.75});
 
   for (auto _ : state) {
-    // These times are in different segments of the fork.
-    fork->Find(Instant() + 1.0 / 3.0 * steps * Second);
-    fork->Find(Instant() + 2.0 / 3.0 * steps * Second);
-    fork->Find(Instant() + 5.0 / 6.0 * steps * Second);
+    // These times are in different segments of the trajectory.
+    trajectory.find(Instant() + 1.0 / 3.0 * steps * Second);
+    trajectory.find(Instant() + 2.0 / 3.0 * steps * Second);
+    trajectory.find(Instant() + 5.0 / 6.0 * steps * Second);
   }
 }
 
 void BM_DiscreteTrajectoryLowerBound(benchmark::State& state) {
+  Instant const t0;
   int const steps = state.range(0);
-  not_null<std::unique_ptr<DiscreteTrajectory<World>>> const trajectory =
-      CreateMotionlessTrajectory(steps);
-  not_null<DiscreteTrajectory<World>*> const fork =
-      ForkAt(*ForkAt(*trajectory, 0.5), 0.75);
+  auto const timeline =
+      NewMotionlessTrajectoryTimeline(World::origin,
+                                      /*Δt=*/1 * Second,
+                                      /*t1=*/t0,
+                                      /*t2=*/t0 + steps * Second);
+  auto const trajectory = MakeTrajectory(timeline, {0.5, 0.75});
 
   for (auto _ : state) {
-    // These times are in different segments of the fork.
-    fork->LowerBound(Instant() + 1.0 / 3.0 * steps * Second);
-    fork->LowerBound(Instant() + 2.0 / 3.0 * steps * Second);
-    fork->LowerBound(Instant() + 5.0 / 6.0 * steps * Second);
+    // These times are in different segments of the trajectory.
+    trajectory.lower_bound(Instant() + 1.0 / 3.0 * steps * Second);
+    trajectory.lower_bound(Instant() + 2.0 / 3.0 * steps * Second);
+    trajectory.lower_bound(Instant() + 5.0 / 6.0 * steps * Second);
   }
 }
 
 void BM_DiscreteTrajectoryEvaluateDegreesOfFreedomExact(
     benchmark::State& state) {
-  int const steps = state.range(0);
-  not_null<std::unique_ptr<DiscreteTrajectory<World>>> const trajectory =
-      CreateCircularTrajectory(4);
+  Instant const t0;
+  auto const timeline =
+      NewCircularTrajectoryTimeline<World>(/*ω=*/1 * Radian / Second,
+                                           /*r=*/1 * Metre,
+                                           /*Δt=*/1 * Second,
+                                           /*t1=*/t0,
+                                           /*t2=*/t0 + 4 * Second);
+  auto const trajectory = MakeTrajectory(timeline, {});
 
   Instant const t = Instant() + 1 * Second;
   for (auto _ : state) {
-    trajectory->EvaluateDegreesOfFreedom(t);
+    trajectory.EvaluateDegreesOfFreedom(t);
   }
 }
 
 void BM_DiscreteTrajectoryEvaluateDegreesOfFreedomInterpolated(
     benchmark::State& state) {
-  not_null<std::unique_ptr<DiscreteTrajectory<World>>> const trajectory =
-      CreateCircularTrajectory(4);
+  Instant const t0;
+  auto const timeline =
+      NewCircularTrajectoryTimeline<World>(/*ω=*/1 * Radian / Second,
+                                           /*r=*/1 * Metre,
+                                           /*Δt=*/1 * Second,
+                                           /*t1=*/t0,
+                                           /*t2=*/t0 + 4 * Second);
+  auto const trajectory = MakeTrajectory(timeline, {});
 
   Instant const t = Instant() + 1.5 * Second;
   for (auto _ : state) {
-    trajectory->EvaluateDegreesOfFreedom(t);
+    trajectory.EvaluateDegreesOfFreedom(t);
   }
 }